requirements.txt
duckdb>=0.9.0
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
scikit-optimize>=0.9.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
tqdm>=4.64.0